collision 수정 (87강)

+++ Gui.cpp
@@ -133,7 +133,7 @@ gui::DropDownList::DropDownList(float x, float y, float width, float height, sf:
                     );
 
 
-  for(size_t i = 0; i < numOfElements; i++) {
+  for(unsigned i = 0; i < numOfElements; i++) {
     this->list.push_back(
                          new gui::Button(
                                          x, y + ((i+1) * height), width, height,
@@ -229,11 +229,11 @@ gui::TextureSelector::TextureSelector(float x, float y, float width, float heigh
   this->sheet.setPosition(x + offset, y);
 
   if(this->sheet.getGlobalBounds().width > this->bounds.getGlobalBounds().width) {
-    this->sheet.setTextureRect(sf::IntRect(0, 0, this->bounds.getGlobalBounds().width, this->sheet.getGlobalBounds().height));
+    this->sheet.setTextureRect(sf::IntRect(0, 0, static_cast<int>(this->bounds.getGlobalBounds().width), static_cast<int>(this->sheet.getGlobalBounds().height)));
   }
 
   if(this->sheet.getGlobalBounds().height > this->bounds.getGlobalBounds().height) {
-    this->sheet.setTextureRect(sf::IntRect(0, 0, this->sheet.getGlobalBounds().width, this->bounds.getGlobalBounds().height));
+    this->sheet.setTextureRect(sf::IntRect(0, 0, static_cast<int>(this->sheet.getGlobalBounds().width), static_cast<int>(this->bounds.getGlobalBounds().height)));
   }
 
   this->selector.setPosition(x + offset, y);
